# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FuelPanda \u2014 Fleet Tracking Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/vehicles": {
      "post": {
        "tags": [
          "vehicles"
        ],
        "summary": "Create Vehicle",
        "operationId": "create_vehicle_api_v1_vehicles_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "vehicles"
        ],
        "summary": "List Vehicles",
        "operationId": "list_vehicles_api_v1_vehicles_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VehicleOut"
                  },
                  "title": "Response List Vehicles Api V1 Vehicles Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vehicles/{vehicle_id}": {
      "put": {
        "tags": [
          "vehicles"
        ],
        "summary": "Update Vehicle",
        "operationId": "update_vehicle_api_v1_vehicles__vehicle_id__put",
        "parameters": [
          {
            "name": "vehicle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Vehicle Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "vehicles"
        ],
        "summary": "Get Vehicle",
        "operationId": "get_vehicle_api_v1_vehicles__vehicle_id__get",
        "parameters": [
          {
            "name": "vehicle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Vehicle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VehicleCreate": {
        "properties": {
          "registration_number": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Registration Number"
          },
          "capacity_gallons": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string"
              }
            ],
            "title": "Capacity Gallons"
          },
          "status": {
            "$ref": "#/components/schemas/VehicleStatus",
            "default": "AVAILABLE"
          }
        },
        "type": "object",
        "required": [
          "registration_number",
          "capacity_gallons"
        ],
        "title": "VehicleCreate"
      },
...
      "VehicleOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "registration_number": {
            "type": "string",
            "title": "Registration Number"
          },
          "capacity_gallons": {
            "type": "string",
            "title": "Capacity Gallons"
          },
          "status": {
            "$ref": "#/components/schemas/VehicleStatus"
          },
          "last_latitude": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Latitude"
          },
          "last_longitude": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Longitude"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "updated_at",
          "registration_number",
          "capacity_gallons",
          "status",
          "last_latitude",
          "last_longitude",
          "last_seen_at"
        ],
        "title": "VehicleOut"
      },
      "VehicleStatus": {
        "type": "string",
        "enum": [
          "AVAILABLE",
          "IN_USE",
          "MAINTENANCE"
        ],
        "title": "VehicleStatus"
      },
      "VehicleUpdate": {
        "properties": {
          "registration_number": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Registration Number"
          },
          "capacity_gallons": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity Gallons"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VehicleStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "VehicleUpdate"
      }
    }
  }
}